Split evaluation function into seperate functions for hnf and nf
Instead of having one function that evaluates a task to normalform,
there are now two seperate functions.
One to evaluate a task to head-normalform and
one that can be used as a wrapper to evaluate a node to normalform.

With this change bfs is no longer complete, however the
time differences between bfs, dfs and fs are now negligable in
deterministic computations.