Improve check for partial calls
Till now a node always had to be locked before
it could be tested for a partial call, because
a different task might be changing a function node
and put it in a state looking like a partial call
for a short time.

Now, when a node is updated, the number of arguments
is set to zero before emptying the children, eliminating
that transitionary state and the need for locking the node.

The check for partial calls is now also included in the
hnf check.