Implement free variable printing. Closes #6
Moved the check of the task result map out of the switch
in the toHnf function and before the return condition.
This way bound free variables will always be recognized.
Also implemented a check of the task result map in the
nf calculation for the same reason.

Furthermore, free variables are now counted and the names are
distinguished based on this counter.