Fix bug in bindIO where the second argument was not evaluated to a partial function call
bindIO only evaluated its first argument to get the result of the
IO action, directly binding the result to the second argument.
This lead to errors when the second argument had not been fully
evaluated to a partial function call, because it would drop the
newly bound result during further evaluation.
Now bindIO returns a call to apply to handle binding of the
result and evaluation of its second argument.