Move external files out of the gocurry directory
The gocurry directory should only have the runtime system files as
it is copied to the go working directory. The external files
are not directory part of the runtime system and are installed
separately outside of the go working directory. Thus
they should not be part of the gocurry directory.