Fix issues with encapsulated search
The tasks used for the search now have a copy of the orginals task fingerprint
in order to properly evaluate potential choice nodes.

Furthermore the results of the search are no longer returned in place
but saved in the task result map of the node starting the search.
This avoids unwanted sharing of results between tasks.