Fix and simplify evaluation of non-linear functional patterns
When =:<= encounters a variable that was marked
in its Task. It replaces the variable and creates a unification.
The functions to test if two variables have been marked before unification
and to find multiple variables in evaluated constructors have been removed.
This also fixes an issue with early replacement of variables.